update the post-save action for preferences to redirect to the user  *if* we were editing a user, otherwise it redirects back the system admin; Resolves http://bugs.web2project.net/view.php?id=856

+++ modules/system/addeditpref.php
@@ -7,7 +7,7 @@ if (!defined('W2P_BASE_DIR')) {
 ## add or edit a user preferences
 ##
 
-$user_id = w2PgetParam($_GET, 'user_id', '0');
+$user_id = (int) w2PgetParam($_GET, 'user_id', '0');
 $perms = &$AppUI->acl();
 // check permissions for this record
 $canEdit = canEdit('system');
